Refactor : 장소 조회 부분 context 구조 변경
장소 조회 부분 context 구조 변경 및 내부 구조 변경
