+++ client/project-there/src/components/journeyDetail/PlaceCategoryPicker/CategoryIcons.jsx
