--- client/project-there/src/contexts/Places.jsx
@@ -1,5 +1,0 @@
-import { createContext } from 'react';
-
-const Places = createContext();
-
-export default Places;
