--- client/project-there/src/pages/PlaceDetailPage.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-function PlaceDetailPage() {
-  return <div />;
-}
-
-export default PlaceDetailPage;
